refactor: consolidate auth utilities
