+++ app/workflow/variables.ts
@@ -14,6 +14,7 @@ export const WORKFLOW_VARIABLES = {
   // Service account
   provisioningUserId: "string",
   provisioningUserEmail: "string",
+  // eslint-disable-next-line sonarjs/no-hardcoded-passwords
   generatedPassword: "string",
 
   // Roles
